+++ examples/Checkbox/CheckboxWithStatus.js
@@ -3,23 +3,9 @@ import React from 'react';
 import Checkbox from 'src/Checkbox';
 import DebugBox from '../DebugBox';
 
-function CheckboxDoc() {
+function CheckboxWithStatusExample() {
     return (
         <div>
-            <h2>&lt;Checkbox&gt;</h2>
-
-            <DebugBox>
-                <Checkbox basic="Count me in" />
-            </DebugBox>
-
-            <DebugBox>
-                <Checkbox
-                    defaultChecked
-                    basic="Join pilot program"
-                    aside="Secondary helps"
-                    tag="New" />
-            </DebugBox>
-
             <DebugBox>
                 <Checkbox
                     indeterminate
@@ -43,16 +29,8 @@ function CheckboxDoc() {
                     status="error"
                     errorMsg="Unauthorized" />
             </DebugBox>
-
-            <DebugBox>
-                <Checkbox
-                    disabled
-                    basic="Join pilot program"
-                    aside="Secondary helps"
-                    tag="New" />
-            </DebugBox>
         </div>
     );
 }
 
-export default CheckboxDoc;
+export default CheckboxWithStatusExample;