+++ examples/Tag/BasicTag.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+import Tag from 'src/Tag';
+
+function BasicTagExample() {
+    return (
+        <Tag>tag</Tag>
+    );
+}
+
+export default BasicTagExample;