+++ examples/Tooltip/TooltipCustomArrowStyle.js
@@ -1,0 +1,13 @@
+import React from 'react';
+
+import Tooltip from 'src/Tooltip';
+
+function TooltipCustomArrowStyleExample() {
+    return (
+        <Tooltip arrowStyle={{ left: '12px' }}>
+            Tooltip
+        </Tooltip>
+    );
+}
+
+export default TooltipCustomArrowStyleExample;